+++ samplesheets/forms.py
@@ -16,6 +16,7 @@ from django.utils import timezone
 from projectroles.models import Project
 from projectroles.plugins import get_backend_api
 
+from samplesheets.constants import HIDDEN_SHEET_TEMPLATE_FIELDS
 from samplesheets.io import SampleSheetIO, ARCHIVE_TYPES
 from samplesheets.utils import clean_sheet_dir_name
 from samplesheets.models import (
@@ -183,6 +184,9 @@ class SheetTemplateCreateForm(forms.Form):
                 self.fields[k] = forms.CharField(**field_kwargs)
                 self.initial[k] = json.dumps(v)
                 self.json_fields.append(k)
+            # Hide fields not intended to be edited (see issue #1443)
+            if k in HIDDEN_SHEET_TEMPLATE_FIELDS:
+                self.fields[k].widget = forms.widgets.HiddenInput()
 
     @classmethod
     def _get_tsv_data(cls, path, file_names):